docs: update utility pages
