+++ docs/app/react/split-props/overview/doc.mdx
