+++ docs/app/react/trap-focus/overview/doc.mdx
@@ -1,4 +1,7 @@
 ---
+heading: 'trapFocus'
+description: 'A utility for trapping focus within a Modal.'
+a11y: 'utilities'
 npm: '@cerberus-design/react'
 source: 'aria-helpers/trap-focus.aria.ts'
 recipe: ''
