--- docs/app/react/trap-focus/page.tsx
@@ -1,21 +1,0 @@
-import ApiLinks from '@/app/components/ApiLinks'
-import OnThisPage from '../../components/OnThisPage'
-import { PageMainContent, PageSections } from '../../components/PageLayout'
-import UseThemeDoc, { frontmatter } from './doc.mdx'
-
-export default function TrapFocusPage() {
-  return (
-    <>
-      <PageMainContent>
-        <ApiLinks {...frontmatter} />
-        <main>
-          <UseThemeDoc />
-        </main>
-      </PageMainContent>
-
-      <PageSections>
-        <OnThisPage />
-      </PageSections>
-    </>
-  )
-}
