+++ docs/app/react/use-modal/overview/page.tsx
@@ -1,0 +1,28 @@
+import ApiLinks from '@/app/components/ApiLinks'
+import OnThisPage from '@/app/components/OnThisPage'
+import { PageMainContent, PageSections } from '@/app/components/PageLayout'
+import UseThemeDoc, { frontmatter } from './doc.mdx'
+import FeatureHeader from '@/app/components/FeatureHeader'
+import type { MatchFeatureKind } from '@/app/components/MatchFeatureImg'
+
+export default function UseModalPage() {
+  return (
+    <>
+      <PageMainContent>
+        <FeatureHeader
+          heading={frontmatter.heading}
+          description={frontmatter.description}
+          a11y={frontmatter.a11y as MatchFeatureKind}
+        />
+        <ApiLinks {...frontmatter} />
+        <main>
+          <UseThemeDoc />
+        </main>
+      </PageMainContent>
+
+      <PageSections>
+        <OnThisPage />
+      </PageSections>
+    </>
+  )
+}
