+++ docs/app/react/use-root-colors/overview/doc.mdx
@@ -1,14 +1,14 @@
 ---
 heading: 'useRootColors'
 description: 'Get Cerberus colors from the document root.'
+a11y: 'utilities'
 npm: '@cerberus-design/react'
 source: 'hooks/useRootColors.ts'
 recipe: ''
-a11y: 'utilities'
 ---
 
 import CodePreview from '@/app/components/CodePreview'
-import CanvasPreview from './components/canvas-preview'
+import CanvasPreview from '@/app/react/use-root-colors/components/canvas-preview'
 
 
 ```ts
