+++ docs/app/react/use-theme-context/overview/doc.mdx
@@ -1,7 +1,9 @@
 ---
+heading: 'useThemeContext'
+description: 'Access or modify the theme and mode properties of the Cerberus Design System in multiple components.'
+a11y: 'utilities'
 npm: '@cerberus-design/react'
 source: 'context/theme.tsx'
-recipe: ''
 ---
 
 import {
