+++ docs/app/react/use-theme/overview/doc.mdx
@@ -1,7 +1,9 @@
 ---
+heading: 'useTheme'
+description: 'Access or modify the theme and mode properties of the Cerberus Design System.'
+a11y: 'utilities'
 npm: '@cerberus-design/react'
 source: 'hooks/useTheme.ts'
-recipe: ''
 ---
 
 import {
