+++ docs/app/react/use-toggle/overview/doc.mdx
@@ -1,7 +1,9 @@
 ---
+heading: 'useToggle'
+description: 'Manage the Toggle state with less code.'
+a11y: 'utilities'
 npm: '@cerberus-design/react'
 source: 'hooks/useToggle.ts'
-recipe: ''
 ---
 
 import {
